Fixed the row displacement
If a title was too long, it would displace the row below it. I fixed
this by ensuring that the titles cannot wrap, and will show an ellipsis
if there was more to the title. While this might not be ideal, it works
reasonably well since the poster art often shows the full title.